pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.14",
    "numpy>=2.3.1",
    "openai>=1.93.0",
    "pydantic>=2.11.7",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
    "tiktoken>=0.9.0",
    "uvicorn>=0.35.0",
]
